+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -216,8 +216,7 @@ internal constructor(
     private val waitConnection = MutableSharedFlow<Result<ConnectionData>>()
     public val clientState: ClientState = mutableClientState
 
-    @InternalStreamChatApi
-    public val streamDateFormatter: StreamDateFormatter = StreamDateFormatter()
+    private val streamDateFormatter: StreamDateFormatter = StreamDateFormatter()
     private val eventsObservable = ChatEventsObservable(waitConnection, userScope, chatSocket)
     private val eventMutex = Mutex()
 
@@ -306,7 +305,7 @@ internal constructor(
         logger.i { "Initialised: ${buildSdkTrackingHeaders()}" }
     }
 
-    private fun handleEvent(event: ChatEvent) {
+    private suspend fun handleEvent(event: ChatEvent) {
         when (event) {
             is ConnectedEvent -> {
                 logger.i { "[handleEvent] event: ConnectedEvent(userId='${event.me.id}')" }
@@ -415,6 +414,7 @@ internal constructor(
 
                 userStateService.onSetUser(user, isAnonymous)
                 chatSocket.connectUser(user, isAnonymous)
+                mutableClientState.setInitializationState(InitializationState.COMPLETE)
                 waitFirstConnection(timeoutMilliseconds)
             }
 
@@ -434,7 +434,10 @@ internal constructor(
                     }
 
                     else -> {
-                        getConnectionId()?.let { Result.Success(ConnectionData(userState.user, it)) }
+                        getConnectionId()?.let {
+                            mutableClientState.setInitializationState(InitializationState.COMPLETE)
+                            Result.Success(ConnectionData(userState.user, it))
+                        }
                             ?: run {
                                 logger.e {
                                     "[setUser] Trying to connect the same user twice without a previously completed " +
@@ -460,8 +463,6 @@ internal constructor(
             }
         }.onErrorSuspend {
             disconnectSuspend(flushPersistence = true)
-        }.onSuccess {
-            mutableClientState.setInitializationState(InitializationState.COMPLETE)
         }
     }
 
@@ -508,6 +509,7 @@ internal constructor(
      *
      * @return [AppSettings] the settings of the app.
      */
+    @CheckResult
     public fun appSettings(): Call<AppSettings> = api.appSettings()
 
     /**
@@ -549,7 +551,7 @@ internal constructor(
         tokenProvider: TokenProvider,
         timeoutMilliseconds: Long?,
     ): Result<ConnectionData> {
-        mutableClientState.setInitializationState(InitializationState.RUNNING)
+        mutableClientState.setInitializationState(InitializationState.INITIALIZING)
         logger.d { "[connectUserSuspend] userId: '${user.id}', username: '${user.name}'" }
         return setUser(user, tokenProvider, timeoutMilliseconds).also { result ->
             logger.v {
@@ -646,9 +648,8 @@ internal constructor(
      * This method initializes [ChatClient] to allow the use of Stream REST API client.
      * Moreover, it warms up the connection, and sets up notifications.
      */
-    @InternalStreamChatApi
-    public suspend fun setUserWithoutConnectingIfNeeded() {
-        if (clientState.initializationState.value == InitializationState.RUNNING) {
+    internal suspend fun setUserWithoutConnectingIfNeeded() {
+        if (clientState.initializationState.value == InitializationState.INITIALIZING) {
             delay(INITIALIZATION_DELAY)
             return setUserWithoutConnectingIfNeeded()
         } else if (isUserSet() || clientState.initializationState.value == InitializationState.COMPLETE) {
@@ -993,21 +994,26 @@ internal constructor(
 
     //endregion
 
-    public fun disconnectSocket() {
-        chatSocket.disconnect()
-    }
-
-    public fun reconnectSocket() {
-        when (val userState = userStateService.state) {
-            is UserState.UserSet, is UserState.AnonymousUserSet -> chatSocket.reconnectUser(
-                userState.userOrError(),
-                userState is UserState.AnonymousUserSet,
-                true,
-            )
-
-            else -> error("Invalid user state $userState without user being set!")
+    @CheckResult
+    public fun disconnectSocket(): Call<Unit> =
+        CoroutineCall(userScope) {
+            Result.Success(chatSocket.disconnect())
         }
-    }
+
+    @CheckResult
+    public fun reconnectSocket(): Call<Unit> =
+        CoroutineCall(userScope) {
+            when (val userState = userStateService.state) {
+                is UserState.UserSet, is UserState.AnonymousUserSet -> Result.Success(
+                    chatSocket.reconnectUser(
+                        userState.userOrError(),
+                        userState is UserState.AnonymousUserSet,
+                        true,
+                    )
+                )
+                else -> Result.Failure(Error.GenericError("Invalid user state $userState without user being set!"))
+            }
+        }
 
     public fun addSocketListener(listener: SocketListener) {
         chatSocket.addListener(listener)
@@ -1456,6 +1462,7 @@ internal constructor(
      *
      * @return Executable async [Call] responsible for sending the Giphy.
      */
+    @CheckResult
     public fun sendGiphy(message: Message): Call<Message> {
         val request = message.run {
             SendActionRequest(cid, id, type, mapOf(KEY_MESSAGE_ACTION to MESSAGE_ACTION_SEND))
@@ -1482,6 +1489,7 @@ internal constructor(
      *
      * @return Executable async [Call] responsible for shuffling the Giphy.
      */
+    @CheckResult
     public fun shuffleGiphy(message: Message): Call<Message> {
         val request = message.run {
             SendActionRequest(cid, id, type, mapOf(KEY_MESSAGE_ACTION to MESSAGE_ACTION_SHUFFLE))
@@ -1556,6 +1564,7 @@ internal constructor(
      *
      * @return Executable async [Call] responsible for sending a message.
      */
+    @CheckResult
     public fun sendMessage(
         channelType: String,
         channelId: String,
@@ -1743,7 +1752,6 @@ internal constructor(
      * @see [queryChannel]
      */
     @CheckResult
-    @InternalStreamChatApi
     private fun queryChannelInternal(
         channelType: String,
         channelId: String,
@@ -2741,6 +2749,7 @@ internal constructor(
      */
     public fun devToken(userId: String): String = tokenUtils.devToken(userId)
 
+    @CheckResult
     internal fun <R, T : Any> Call<T>.precondition(
         pluginsList: List<R>,
         preconditionCheck: suspend R.() -> Result<Unit>,