+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/EventMapping.kt
@@ -588,6 +588,7 @@ private fun ChannelUserBannedEventDto.toDomain(): ChannelUserBannedEvent {
         channelId = channel_id,
         user = user.toDomain(),
         expiration = expiration,
+        shadow = shadow ?: false,
     )
 }