+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/TextUtils.kt
@@ -17,6 +17,7 @@
 package io.getstream.chat.android.compose.ui.util
 
 import android.annotation.SuppressLint
+import android.text.util.Linkify
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.AnnotatedString
@@ -25,6 +26,7 @@ import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.core.util.PatternsCompat
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
+import java.util.regex.Pattern
 
 /**
  * Takes the given message text and builds an annotated message text that shows links and allows for clicks,
@@ -36,6 +38,7 @@ import io.getstream.chat.android.compose.ui.theme.ChatTheme
  * @return The annotated String, with clickable links, if applicable.
  */
 @Composable
+@SuppressLint("RestrictedApi")
 internal fun buildAnnotatedMessageText(text: String, color: Color): AnnotatedString {
     return buildAnnotatedString {
         // First we add the whole text to the [AnnotatedString] and style it as a regular text.
@@ -51,38 +54,71 @@ internal fun buildAnnotatedMessageText(text: String, color: Color): AnnotatedStr
 
         // Then for each available link in the text, we add a different style, to represent the links,
         // as well as add a String annotation to it. This gives us the ability to open the URL on click.
-        @SuppressLint("RestrictedApi")
-        val matcher = PatternsCompat.AUTOLINK_WEB_URL.matcher(text)
-        while (matcher.find()) {
-            val start = matcher.start()
-            val end = matcher.end()
+        linkify(
+            text = text,
+            tag = "URL",
+            pattern = PatternsCompat.AUTOLINK_WEB_URL,
+            matchFilter = Linkify.sUrlMatchFilter,
+            schemes = URL_SCHEMES,
+            linkColor = ChatTheme.colors.primaryAccent,
+        )
+        linkify(
+            text = text,
+            tag = "EMAIL",
+            pattern = PatternsCompat.AUTOLINK_EMAIL_ADDRESS,
+            schemes = EMAIL_SCHEMES,
+            linkColor = ChatTheme.colors.primaryAccent,
+        )
+    }
+}
 
-            addStyle(
-                style = SpanStyle(
-                    color = ChatTheme.colors.primaryAccent,
-                    textDecoration = TextDecoration.Underline,
-                ),
-                start = start,
-                end = end,
-            )
+private fun AnnotatedString.Builder.linkify(
+    text: CharSequence,
+    tag: String,
+    pattern: Pattern,
+    matchFilter: Linkify.MatchFilter? = null,
+    schemes: List<String>,
+    linkColor: Color
+) {
+    @SuppressLint("RestrictedApi")
+    val matcher = pattern.matcher(text)
+    while (matcher.find()) {
+        val start = matcher.start()
+        val end = matcher.end()
 
-            val linkText = requireNotNull(matcher.group(0)!!)
-
-            // Add "http://" prefix if link has no scheme in it
-            val url = if (URL_SCHEMES.none { scheme -> linkText.startsWith(scheme) }) {
-                URL_SCHEMES[0] + linkText
-            } else {
-                linkText
-            }
-
-            addStringAnnotation(
-                tag = "URL",
-                annotation = url,
-                start = start,
-                end = end,
-            )
+        if (matchFilter != null && !matchFilter.acceptMatch(text, start, end)) {
+            continue
         }
+
+        addStyle(
+            style = SpanStyle(
+                color = linkColor,
+                textDecoration = TextDecoration.Underline,
+            ),
+            start = start,
+            end = end,
+        )
+
+        val linkText = requireNotNull(matcher.group(0)!!)
+
+        val url = linkText.fixPrefix(schemes)
+
+        addStringAnnotation(
+            tag = tag,
+            annotation = url,
+            start = start,
+            end = end,
+        )
+    }
+}
+
+private fun String.fixPrefix(schemes: List<String>): String {
+    return if (schemes.none { scheme -> startsWith(scheme) }) {
+        schemes[0] + this
+    } else {
+        this
     }
 }
 
 private val URL_SCHEMES = listOf("http://", "https://")
+private val EMAIL_SCHEMES = listOf("mailto:")